🚚  Move moderation files to their own folder
